backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI Integration
google-generativeai==0.8.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Monitoring & Performance
psutil==5.9.6

# Caching & Compression
python-lru==1.0.3

# Logging
python-json-logger==2.0.7

# Development tools (optional)
# black==23.11.0
# isort==5.12.0
# pytest==7.4.3
# pytest-asyncio==0.21.1